roachpb: remove settings from package
`pb` packages are supposed to have as little dependencies as possible,
they should not depend on higher-level machinery like `settings`.
This commit moves the span stats settings out of `roachpb`. This will
prevent a circular dependency if we move the cluster version setting
code to `settings`.

Epic: none
Release note: None
