+++ pkg/roachpb/BUILD.bazel
@@ -15,7 +15,6 @@ go_library(
         "metadata_replicas.go",
         "span_config.go",
         "span_group.go",
-        "span_stats.go",
         "tenant.go",
         "version.go",
     ],
@@ -29,7 +28,6 @@ go_library(
         "//pkg/kv/kvserver/allocator/load",
         "//pkg/kv/kvserver/concurrency/isolation",
         "//pkg/kv/kvserver/concurrency/lock",
-        "//pkg/settings",
         "//pkg/storage/enginepb",
         "//pkg/util",
         "//pkg/util/admission/admissionpb",
